refactor: Rename `lib/plugins/aws/info/getApiKeyValues.js`
