+++ lib/plugins/aws/info/index.js
@@ -6,7 +6,7 @@ const writeServiceOutputs = require('../../../cli/write-service-outputs');
 const validate = require('../lib/validate');
 const getStackInfo = require('./getStackInfo');
 const getResourceCount = require('./getResourceCount');
-const getApiKeyValues = require('./getApiKeyValues');
+const getApiKeyValues = require('./get-api-key-values');
 const display = require('./display');
 const { log } = require('@serverless/utils/log');
