+++ test/unit/lib/plugins/aws/info/get-api-key-values.test.js
